+++ application/shared/view/timeclock/admin/jobs/home.php
@@ -32,6 +32,7 @@
                             <th>Job ID</th>
                             <th>Client Name</th>
                             <th>Job Name</th>
+                            <th>Quoted Time</th>
                             <th>Status</th>
                             <th>Manage</th>
                         </tr>
@@ -58,6 +59,7 @@
                             <td onclick="jobTableClicked('view', '{jobs[<?php echo $i; ?>]['job_uid']}')">{jobs[<?php echo $i; ?>]['job_uid']}</td>
                             <td onclick="jobTableClicked('view', '{jobs[<?php echo $i; ?>]['job_uid']}')">{jobs[<?php echo $i; ?>]['client_name']}</td>
                             <td onclick="jobTableClicked('view', '{jobs[<?php echo $i; ?>]['job_uid']}')">{jobs[<?php echo $i; ?>]['job_name']}</td>
+                            <td onclick="jobTableClicked('view', '{jobs[<?php echo $i; ?>]['job_uid']}')"><?php echo $this->model_jobs->quoted_hours($this->sys->template->jobs[$i]['job_id']); ?></td>
                             <td onclick="jobTableClicked('view', '{jobs[<?php echo $i; ?>]['job_uid']}')"><div class="<?php echo $status[0]; ?>"><?php echo $status[1]; ?></div></td>
                                 <td>
                                     <ul class="icons">
@@ -75,6 +77,7 @@
                             <th>Job ID</th>
                             <th>Client Name</th>
                             <th>Job Name</th>
+                            <th>Quoted Time</th>
                             <th>Status</th>
                             <th>Manage</th>
                         </tr>