+++ application/shared/view/timeclock/jobs/home.php
@@ -22,6 +22,7 @@
                             <th>Job ID</th>
                             <th>Client Name</th>
                             <th>Job Name</th>
+                            <th>Quoted Time</th>
                             <th>Status</th>
                         </tr>
                     </thead>
@@ -47,6 +48,7 @@
                             <td onclick="jobTableClicked('view', '{jobs[<?php echo $i; ?>]['job_uid']}')">{jobs[<?php echo $i; ?>]['job_uid']}</td>
                             <td onclick="jobTableClicked('view', '{jobs[<?php echo $i; ?>]['job_uid']}')">{jobs[<?php echo $i; ?>]['client_name']}</td>
                             <td onclick="jobTableClicked('view', '{jobs[<?php echo $i; ?>]['job_uid']}')">{jobs[<?php echo $i; ?>]['job_name']}</td>
+                            <td onclick="jobTableClicked('view', '{jobs[<?php echo $i; ?>]['job_uid']}')"><?php echo $this->model_jobs->quoted_hours($this->sys->template->jobs[$i]['job_id']); ?></td>
                             <td onclick="jobTableClicked('view', '{jobs[<?php echo $i; ?>]['job_uid']}')"><div class="<?php echo $status[0]; ?>"><?php echo $status[1]; ?></div></td>
                             </tr>
                             <?php
@@ -58,6 +60,7 @@
                             <th>Job ID</th>
                             <th>Client Name</th>
                             <th>Job Name</th>
+                            <th>Quoted Time</th>
                             <th>Status</th>
                         </tr>
                     </tfoot>